Refactoring of regression
